+++ src/components/Fieldset/Fieldset.test.tsx
@@ -39,17 +39,7 @@ describe("Fieldset", () => {
     expect(screen.getByText("Some other fields")).toBeInTheDocument();
   });
 
-  it("renders 'Optional' text by default in the legend", () => {
-    render(
-      <Fieldset legendText="Legend Text">
-        <p>Some other fields</p>
-      </Fieldset>
-    );
-
-    expect(screen.getByText(/Optional/i)).toBeInTheDocument();
-  });
-
-  it("renders 'Required' text in the legend", () => {
+  it("renders '(Required)' text in the legend when `isRequired` is true", () => {
     render(
       <Fieldset legendText="Legend Text" isRequired>
         <p>Some other fields</p>
@@ -59,7 +49,7 @@ describe("Fieldset", () => {
     expect(screen.getByText(/Required/i)).toBeInTheDocument();
   });
 
-  it("can hide the 'Required'/'Optional' text in the legend but still show the main text", () => {
+  it("can hide the '(Required)' text in the legend but still show the main text", () => {
     const { rerender } = render(
       <Fieldset legendText="Legend Text" isRequired>
         <p>Some other fields</p>
@@ -69,16 +59,7 @@ describe("Fieldset", () => {
     expect(screen.getByText(/Required/i)).toBeInTheDocument();
 
     rerender(
-      <Fieldset legendText="Legend Text" optReqFlag={false}>
-        <p>Some other fields</p>
-      </Fieldset>
-    );
-
-    expect(screen.getByText(/Legend Text/i)).toBeInTheDocument();
-    expect(screen.queryByText(/Optional/i)).not.toBeInTheDocument();
-
-    rerender(
-      <Fieldset legendText="Legend Text" isRequired optReqFlag={false}>
+      <Fieldset legendText="Legend Text" isRequired showRequiredLabel={false}>
         <p>Some other fields</p>
       </Fieldset>
     );
@@ -104,7 +85,7 @@ describe("Fieldset", () => {
       .toJSON();
     const hiddenHelperText = renderer
       .create(
-        <Fieldset legendText="Legend Text" isRequired optReqFlag={false}>
+        <Fieldset legendText="Legend Text" isRequired showRequiredLabel={false}>
           <p>Some other fields</p>
         </Fieldset>
       )