+++ src/components/RadioGroup/RadioGroup.tsx
@@ -44,13 +44,14 @@ export interface RadioGroupProps {
   name: string;
   /** The action to perform on the `<input>`'s onChange function  */
   onChange?: (value: string) => void;
-  /** Whether or not to display "Required"/"Optional" in the label text. */
-  optReqFlag?: boolean;
   /** Offers the ability to hide the helper/invalid text. */
   showHelperInvalidText?: boolean;
   /** Offers the ability to show the group's legend onscreen or hide it. Refer
    * to the `labelText` property for more information. */
   showLabel?: boolean;
+  /** Whether or not to display the "(Required)" text in the label text.
+   * True by default. */
+  showRequiredLabel?: boolean;
 }
 
 const noop = () => {};
@@ -77,9 +78,9 @@ const RadioGroup = React.forwardRef<
     layout = LayoutTypes.Column,
     name,
     onChange = onChangeDefault,
-    optReqFlag = true,
     showHelperInvalidText = true,
     showLabel = true,
+    showRequiredLabel = true,
   } = props;
   const footnote: HelperErrorTextType = isInvalid ? invalidText : helperText;
   const spacingProp = layout === LayoutTypes.Column ? spacing.s : spacing.l;
@@ -131,8 +132,9 @@ const RadioGroup = React.forwardRef<
       className={className}
       id={`radio-group-${id}`}
       isLegendHidden={!showLabel}
+      isRequired={isRequired}
       legendText={labelText}
-      optReqFlag={optReqFlag}
+      showRequiredLabel={showRequiredLabel}
     >
       <Stack
         aria-label={!showLabel ? labelText : null}