+++ src/components/Select/Select.tsx
@@ -51,8 +51,9 @@ export interface SelectProps {
   /** Offers the ability to show the select's label onscreen or hide it. Refer
    * to the `labelText` property for more information. */
   showLabel?: boolean;
-  /** Whether or not to display the "Required"/"Optional" text in the label text. */
-  showOptReqLabel?: boolean;
+  /** Whether or not to display the "(Required)" text in the label text.
+   * True by default. */
+  showRequiredLabel?: boolean;
   /** The value of the selected option.
    * Should be passed along with `onChange` for controlled components. */
   value?: string;
@@ -82,12 +83,11 @@ const Select = React.forwardRef<
     placeholder,
     showHelperInvalidText = true,
     showLabel = true,
-    showOptReqLabel = true,
+    showRequiredLabel = true,
     selectType = SelectTypes.Default,
     value = "",
   } = props;
   const ariaAttributes = {};
-  const optReqFlag = isRequired ? "Required" : "Optional";
   const styles = useMultiStyleConfig("CustomSelect", { variant: selectType });
   const finalInvalidText = invalidText
     ? invalidText
@@ -112,7 +112,7 @@ const Select = React.forwardRef<
         <Label
           id={`${id}-label`}
           htmlFor={id}
-          optReqFlag={showOptReqLabel && optReqFlag}
+          isRequired={showRequiredLabel && isRequired}
         >
           {labelText}
         </Label>